requirements-dev.txt
pandas
numpy
scikit-learn
pyarrow
joblib
pyyaml
pytest
ruff